test: extract common tests for fs-serve
